Refactored multiple heap-related files by removing unnecessary comments and descriptions for improved clarity. Added new implementations for various problems including Design Twitter, Find Median from Data Stream, Hand of Straights, K Closest Points to Origin, Kth Largest Element in an Array, Last Stone Weight, Merge k Sorted Lists, Minimum Interval to Include Each Query, and Task Scheduler.

+++ Heaps/Find_Median_from_Data_Stream.cpp
@@ -1,16 +1,20 @@
+// https://leetcode.com/problems/find-median-from-data-stream/
+
 #include<bits/stdc++.h>
 using namespace std;
 
 class MedianFinder {
-    priority_queue<long> small, large;
+    priority_queue<int> small, large;
 public:
-    MedianFinder() {}
+    MedianFinder() {
+        
+    }
     
     void addNum(int num) {
         small.push(num);
         large.push(-small.top());
         small.pop();
-        if (small.size() < large.size()) {
+        if(small.size() < large.size()) {
             small.push(-large.top());
             large.pop();
         }
@@ -18,7 +22,7 @@ public:
     
     double findMedian() {
         if(small.size() == large.size())
-            return (small.top() - large.top()) / 2.0;
+            return ((double)small.top() - (double)large.top()) / 2.0;
         return small.top();
     }
 };
